Only generate Tasks for workitems w/ 'task' param
Only expressions in the form of `role :task => 'task_name'` (that fall
into the storage) should be converted to Tasks when using the collection
init methods.  Also removed `to_task` lambda in favor of
`#from_workitems` - having a method called `to_task` on Task that
actually generated a lambda that ostensibly operates on a workitem, not
a task, is very confusing.  Also this way we can reject at the time of
workitem lookup, rather than after instantiating Tasks.