Move all_files to support (and fix yield order)
We already have a support library, so things like `all_files` should be
there (since it's not only relevant to the base Bumbleworks module..
further refactoring forthcoming).  Also, fixed the yield order - since
the method is aggregating files from a directory, the expected result is
pathnames (which a Dir glob would give us).. the name is extra candy.
This also enables us to only use the first param in block params if we
just need the path.