Fix deprecation warning
RSpec doesn't want us to expect a specific error to not be raised.
This actually makes a lot of sense, since something of a false
positive could occur if the method goes wrong in some unexpected way,
raising a different error than the one "not expected" - which would
make the test pass.  What we really care about is that *no* error
is raised.