Add Broadcast/Wait for Event expressions to Ruote
Our first custom expressions - these are basically cleaner ways of
setting up a specific type of :tag & await pair.  In vanilla route,
it's a bit of a pain to simply broadcast an event (leave a tag) and
catch that event globally with a :where clause without overwriting
any workitem data in the receiver.  This first iteration can only
compare sent workitem to received workitem, using the "this:" and
"event:" dollar prefixes; later we'll implement sending arbitrary
data with the broadcast that can be assessed in the where clause.