Ability to configure options for storage
Added a new configuration setting - 'storage_options' - which defaults
to an empty hash and will be sent as the second argument to the
new_storage method on the adapter.  The base adapter and Hash
adapter (for testing) ignore the options hash completely, but it
can be handy for adapters such as the Redis and Sequel ones.