Add task filtering by process
This is significant because it changes the lookup
method used in the Finder class.  Note that we're
now skipping the storage participant and going
straight to the storage, so we can take advantage
of filtering by wfid directly in the storage
query.  An even more significant LoD violation
than before, but given near-future goals, one
I feel okay with for the time being.