Resimplify Participant classes, more renames
Keep the participants in their own directory; that was a good idea.  The
reorg, not so much.  Going back to the simpler hierarchy, but the
error handler participant has been renamed to the ErrorDispatcher to
better differentiate it from the ErrorHandler interface class
(implementations of which are *subscribers* to the dispatcher).