Add entity lookup to Bumbleworks::Process
This is a little trickier than Task or Participant entity lookup, since
a running process may have multiple workitems (in multiple active
leaves of the workflow).  The design here is that if all leaves agree
on which entity is represented, that entity will be returned - otherwise
an EntityConflict exception will be raised.