Task finder - add with_fields and #where shortcut
Using Task::Finder#where lets you specify several options in the same
method call - this is just a shortcut and does the same thing as
chaining the specified methods.  In fact, it can be combined with method
chaining.  The #completable query also gained an inverse option (send
false as an argument).