Specify extra fields/variables on process launch
Now, when calling #launch_process on a Bumbleworks::Entity, you can
specify additional fields and variables to be sent to the launch
method - these will be merged with the return values from process_fields
and process_variables.  Also, refactored the entity module, mostly
insignificant, but the important change is that :column is replaced with
:attribute, so if you override the persistence attribute, use that key
instead.  This is done to reflect a lack of dependency on any specific
form of persistence.