Optionally leave PIDs alone when cancelling all
When cancelling all processes on an Bumbleworks::Entity, by default all
process identifiers are cleared (by persisting nil to them).  This is
undesirable if the entity update a) leads to side effects or b) is
performed when the entity is being destroyed anyway.  Now you can send
a :clear_identifiers option to cancel_all_processes! - it defaults to
true, but if you explicitly set it to false, the pids will be left
alone.