Change root discovery and default directory logic
It used to be that the default directories included the assumed lib
path (lib/bumbleworks), which meant that the default Bumbleworks.root
was the base app directory.  This made it difficult to set the root
directory somewhere else without either replicating the lib/bumbleworks
hierarchy (for tasks, participants, and processes directories), or
explicitly setting those paths in config as well.  Now, the default root
itself includes lib/bumbleworks (appended to the framework's app root),
which makes it much easier to relocate your BW directories.