Only use brute force process kill when :force true
Just clearing the storage (or removing all of a process's messages) can
result in the worker losing its mind, if a worker is running and trying
to process messages, trackers, etc while the storage is emptied.  If you
know it's safe to do so, you can add :force => true to the .kill_process
and .kill_all_processes methods.