Add magic entity-specific tags for broadcast/wait
Use of the entity-specific tag feature (with :for_entity => true on the
expression) dramatically improves efficiency, especially with a high
number of trackers waiting for the same tag, but only where entities
match.  Note that this doesn't completely replace the functionality of
the :where => :entities_match clause, since the latter allows for
global tags that can trigger both a global listener and an entity-
specific listener.