Add Expression and Process::ErrorRecord wrappers
An Expression instance wraps a FlowExpression, for a more logical API.
A Process::ErrorRecord instance does the same for Ruote::ProcessError
instances, also cleaning up the error message and adding the ability
to #reify the error (re-instantiate the original error).  The Process
class has been changed accordingly, to return the new classes for
relevant queries.